+++ src/main/java/kitchenpos/application/TableGroupService.java
@@ -3,10 +3,10 @@ package kitchenpos.application;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.OrderTable;
-import kitchenpos.domain.Orders;
-import kitchenpos.domain.TableGroup;
-import kitchenpos.domain.Tables;
+import kitchenpos.domain.order.OrderTable;
+import kitchenpos.domain.order.Orders;
+import kitchenpos.domain.order.TableGroup;
+import kitchenpos.domain.order.Tables;
 import kitchenpos.dto.OrderTableRequest;
 import kitchenpos.dto.OrderTableResponse;
 import kitchenpos.dto.TableGroupRequest;
