+++ src/main/java/kitchenpos/domain/order/Orders.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.List;
