+++ src/main/java/kitchenpos/domain/order/Tables.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.order;
 
 import java.util.List;
 import java.util.Objects;
