+++ src/main/java/kitchenpos/repository/MenuProductRepository.java
@@ -1,7 +1,7 @@
 package kitchenpos.repository;
 
 import java.util.List;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.MenuProduct;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
