+++ src/main/java/kitchenpos/repository/OrderRepository.java
@@ -1,8 +1,8 @@
 package kitchenpos.repository;
 
 import java.util.List;
-import kitchenpos.domain.Order;
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.order.Order;
+import kitchenpos.domain.order.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
